requirements.txt
# Bull BC1 - Bitcoin Bull Detection Bot (Long-Only)
# Requirements file

# Core
python-dotenv>=1.0.0
aiohttp>=3.9.0
websockets>=12.0
pyyaml>=6.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Machine Learning
tensorflow>=2.15.0
xgboost>=2.0.0
scikit-learn>=1.3.0

# Hyperparameter Tuning (optional)
optuna>=3.5.0

# Utilities
pytz>=2023.3

# MetaTrader 5 (Windows only)
# MetaTrader5>=5.0.45

# Development & Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
